add current endpoint

+++ src/api/common.ts
@@ -1,6 +1,11 @@
 import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
 
-const baseUrl = 'http://localhost:3004';
+import { getLocationSearch } from '../utils/locationSearch';
+
+const { userId } = getLocationSearch();
+
+const isDev = process.env.NODE_ENV === 'development';
+const baseUrl = 'https://telegram-crm.rossko.dev';
 
 class ApiClient {
     private api: AxiosInstance;
@@ -8,7 +13,7 @@ class ApiClient {
     constructor() {
         this.api = axios.create({
             baseURL: baseUrl,
-            timeout: 3000,
+            timeout: 6000,
             withCredentials: true,
             headers: {
                 'Content-Type': 'application/json',
@@ -17,19 +22,19 @@ class ApiClient {
     }
 
     get(url: string) {
-        return this.api.get(url);
+        return this.api.get(`${baseUrl}/api/${url}?userId=${isDev ? 51673 : userId}`);
     }
 
     post(url: string, { data }: AxiosRequestConfig) {
-        return this.api.post(url, data);
+        return this.api.post(`${baseUrl}/api/${url}?userId=${isDev ? 51673 : userId}`, data);
     }
 
     put(url: string, { data }: AxiosRequestConfig) {
-        return this.api.put(url, data);
+        return this.api.put(`${baseUrl}/api/${url}?userId=${isDev ? 51673 : userId}`, data);
     }
 
     delete(url: string) {
-        return this.api.delete(url);
+        return this.api.delete(`${baseUrl}/api/${url}?userId=${isDev ? 51673 : userId}`);
     }
 }
